+++ packages/atomic/src/components/atomic-search-interface/atomic-search-interface.tsx
@@ -88,6 +88,11 @@ export class AtomicSearchInterface {
    */
   @Prop() public reflectStateInUrl = true;
 
+  /**
+   * The CSS selector for the container where the interface will scroll back to.
+   */
+  @Prop() public scrollContainer = 'atomic-search-interface';
+
   public constructor() {
     setCoveoGlobal();
   }
@@ -149,6 +154,19 @@ export class AtomicSearchInterface {
     }
 
     this.hangingComponentsInitialization.push(event);
+  }
+
+  @Listen('atomic/scrollToTop')
+  public scrollToTop() {
+    const scrollContainerElement = document.querySelector(this.scrollContainer);
+    if (!scrollContainerElement) {
+      this.bindings.engine.logger.warn(
+        `Could not find the scroll container with the selector "${this.scrollContainer}". This will prevent UX interactions that require a scroll from working correctly. Please check the CSS selector in the scrollContainer option`
+      );
+      return;
+    }
+
+    scrollContainerElement.scrollIntoView({behavior: 'smooth'});
   }
 
   /**