+++ packages/headless/src/features/triggers/triggers-slice.ts
@@ -2,12 +2,10 @@ import {createReducer} from '@reduxjs/toolkit';
 import {executeSearch} from '../search/search-actions';
 import {getTriggerInitialState} from './triggers-state';
 import {
-  Trigger,
-  TriggerRedirect,
   isRedirectTrigger,
   isQueryTrigger,
-  TriggerQuery,
   isNotifyTrigger,
+  Trigger,
   TriggerNotify,
 } from './../../api/search/trigger';
 
@@ -15,19 +13,18 @@ export const triggerReducer = createReducer(
   getTriggerInitialState(),
   (builder) =>
     builder.addCase(executeSearch.fulfilled, (state, action) => {
-      const redirectTriggers: Trigger[] = action.payload.response.triggers.filter(
-        (trigger) => isRedirectTrigger(trigger)
+      const redirectTriggers = action.payload.response.triggers.filter(
+        isRedirectTrigger
       );
       state.redirectTo = redirectTriggers.length
-        ? (redirectTriggers[0] as TriggerRedirect).content
+        ? redirectTriggers[0].content
         : '';
 
-      const queryTriggers: Trigger[] = action.payload.response.triggers.filter(
-        (trigger) => isQueryTrigger(trigger)
+      const queryTriggers = action.payload.response.triggers.filter(
+        isQueryTrigger
       );
-      state.query = queryTriggers.length
-        ? (queryTriggers[0] as TriggerQuery).content
-        : '';
+      state.query = queryTriggers.length ? queryTriggers[0].content : '';
+
       const notifyTriggers: Trigger[] = action.payload.response.triggers.filter(
         (trigger) => isNotifyTrigger(trigger)
       );