+++ packages/samples/headless-react/src/components/triggers/query-trigger.class.tsx
@@ -32,11 +32,14 @@ export class QueryTrigger extends Component<{}, QueryTriggerState> {
     if (!this.state) {
       return null;
     }
-    return (
-      <div>
-        The query changed from {this.state.prevQuery + ' '}
-        to {this.state.newQuery}
-      </div>
-    );
+    if (this.state.wasQueryModified) {
+      return (
+        <div>
+          The query changed from {this.state.originalQuery + ' '}
+          to {this.state.newQuery}
+        </div>
+      );
+    }
+    return null;
   }
 }